# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "maeum-pet-store",
    "description": "Delicious fastify.js boilerplate",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
...
      "ITagDto": {
        "title": "ITagDto",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "category id"
          },
          "name": {
            "type": "string",
            "description": "category name",
            "maxLength": 200,
            "minLength": 3
          }
        },
        "required": [
          "id",
          "name"
        ],
        "description": "TagDto of Pet, Store"
      },
...
      "IPostTagBodyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "category name",
            "maxLength": 200,
            "minLength": 3
          }
        },
        "required": [
          "name"
        ],
        "title": "IPostTagBodyDto"
      },
...
      "IPutTagBodyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tag name",
            "maxLength": 200,
            "minLength": 3
          }
        },
        "required": [
          "name"
        ],
        "title": "IPutTagBodyDto"
      },
...
    }
  },
  "paths": {
...
    "/tag/{id}": {
      "delete": {
        "operationId": "delete-tag-by-id",
        "summary": "Delete Tag",
        "tags": [
          "Tag"
        ],
        "description": "Delete Tag using id",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "6f1a29b5-cdc6-4023-a63f-7696d963d09a"
            },
            "in": "query",
            "name": "tid",
            "required": true,
            "description": "request id from client"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "tag id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ITagDto"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    },
                    "validation": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "data": {},
                          "schemaPath": {
                            "type": "string"
                          },
                          "params": {}
                        },
                        "required": [
                          "message",
                          "schemaPath"
                        ]
                      },
                      "description": "information of validator result"
                    }
                  },
                  "required": [
                    "message",
                    "validation"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "read-tag",
        "summary": "Read Tag",
        "tags": [
          "Tag"
        ],
        "description": "Read Tag using id",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "6f1a29b5-cdc6-4023-a63f-7696d963d09a"
            },
            "in": "query",
            "name": "tid",
            "required": true,
            "description": "request id from client"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "tag id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ITagDto"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    },
                    "validation": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "data": {},
                          "schemaPath": {
                            "type": "string"
                          },
                          "params": {}
                        },
                        "required": [
                          "message",
                          "schemaPath"
                        ]
                      },
                      "description": "information of validator result"
                    }
                  },
                  "required": [
                    "message",
                    "validation"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "create-tag",
        "summary": "Create Tag",
        "tags": [
          "Tag"
        ],
        "description": "Create Tag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IPostTagBodyDto"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "6f1a29b5-cdc6-4023-a63f-7696d963d09a"
            },
            "in": "query",
            "name": "tid",
            "required": true,
            "description": "request id from client"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ITagDto"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    },
                    "validation": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "data": {},
                          "schemaPath": {
                            "type": "string"
                          },
                          "params": {}
                        },
                        "required": [
                          "message",
                          "schemaPath"
                        ]
                      },
                      "description": "information of validator result"
                    }
                  },
                  "required": [
                    "message",
                    "validation"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "put-tag",
        "summary": "Put Tag",
        "tags": [
          "Tag"
        ],
        "description": "Put Tag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IPutTagBodyDto"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "6f1a29b5-cdc6-4023-a63f-7696d963d09a"
            },
            "in": "query",
            "name": "tid",
            "required": true,
            "description": "request id from client"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ITagDto"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    },
                    "validation": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "data": {},
                          "schemaPath": {
                            "type": "string"
                          },
                          "params": {}
                        },
                        "required": [
                          "message",
                          "schemaPath"
                        ]
                      },
                      "description": "information of validator result"
                    }
                  },
                  "required": [
                    "message",
                    "validation"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "code of error"
                    },
                    "message": {
                      "type": "string",
                      "description": "description of error"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    }
  }
}